move tests related to functions in _backend to test_backend
